Refactor pdf package; split into cluster format and cluster info getter files
